Fix crash
In Inkscape 0.91, unittouu / uutounit have been moved to the inkex.Effect class, see Release notes: http://wiki.inkscape.org/wiki/index.php/Release_notes/0.91 . It takes a unit in the call to the function, too, but I think in the case of this extension, that can be left out (fallback is px). A user can resize a rendered chart as needed easily.